Add API to return the subscription status
