+++ subscriptions/subscriptions-api/src/main/java/com/duckduckgo/subscriptions/api/Subscriptions.kt
@@ -42,10 +42,25 @@ interface Subscriptions {
      * @return `true` if the Privacy Pro product is available for the user, `false` otherwise
      */
     suspend fun isEligible(): Boolean
+
+    /**
+     * @return `SubscriptionStatus` with the current subscription status
+     */
+    suspend fun getSubscriptionStatus(): SubscriptionStatus
 }
 
 enum class Product(val value: String) {
     NetP("Network Protection"),
     ITR("Identity Theft Restoration"),
     PIR("Data Broker Protection"),
+}
+
+enum class SubscriptionStatus(val statusName: String) {
+    AUTO_RENEWABLE("Auto-Renewable"),
+    NOT_AUTO_RENEWABLE("Not Auto-Renewable"),
+    GRACE_PERIOD("Grace Period"),
+    INACTIVE("Inactive"),
+    EXPIRED("Expired"),
+    UNKNOWN("Unknown"),
+    WAITING("Waiting"),
 }
